refactor(tests): consolidate data map deserialization tests
Merge streaming_private_data_map.rs and streaming_public_data_map.rs into
single streaming_data_map_deserialization.rs file to reduce code duplication.

Key improvements:
- Shared helper functions for chunk info validation
- Single deserialize_data_map function for both private/public use cases
- Consolidated test utilities (create_test_chunk_info, serialize_to_bytes)
- All functions return eyre::Result instead of using unwrap/expect
- 11 comprehensive tests covering both private and public scenarios

Tests verify backward compatibility between old (DataMapLevel) and new
(DataMap) formats for data map deserialization used in streaming downloads.
